requirements_hyperopt.txt
# Hyperparameter Optimization Dependencies
optuna>=3.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0

# Optional: For advanced visualizations
plotly>=5.0.0
kaleido>=0.2.1  # For static plot export

# Optional: For parallel optimization
joblib>=1.1.0

# Optional: For database storage of trials
sqlite3  # Usually included with Python 